requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
python-dotenv>=1.0.0
